refactor query_scorer for simple raw_scorer impl
