+++ lib/segment/src/vector_storage/mod.rs
@@ -19,6 +19,7 @@ mod tests;
 mod async_io;
 mod async_io_mock;
 pub mod common;
+mod query_scorer;
 
 pub use raw_scorer::*;
 pub use vector_storage_base::*;
