+++ lib/segment/src/vector_storage/query_scorer/query_scorer.rs
